requirements-dev.txt
# Development and test dependencies for the Energy Compliance Hub
# Install alongside app/requirements.txt:
#   pip install -r app/requirements.txt -r requirements-dev.txt

pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
anyio>=4.0.0

# App runtime deps needed for import during tests
fastapi>=0.110.0
uvicorn>=0.27.0
sse-starlette>=1.6.0
pandas>=2.0.0

# Mocking / stubs for external dependencies that the app imports at module level
mlflow>=2.15
openai>=1.10.0

# databricks-sdk is required by server/config.py at import time; the tests mock
# the actual network calls, but the module still needs to be importable.
databricks-sdk>=0.20.0
